pyproject.toml
[project]
name = "mlops-recruitair-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lambda-layer-kubectl-v33>=2.0.0",
    "aws-cdk-lib==2.215.0",
    "constructs>=10.0.0,<11.0.0",
    "pyyaml>=6.0.3",
]
